+++ RType/Client/Includes/SoundFactory.hpp
@@ -5,7 +5,7 @@
 // Login   <Volto@epitech.net>
 // 
 // Started on  Fri Nov 20 13:22:47 2015 Probola
-// Last update Tue Dec 15 15:00:31 2015 Probola
+// Last update Tue Dec 15 17:59:36 2015 Probola
 //
 
 #ifndef					__SOUND_FACTORY_HPP__
@@ -21,12 +21,12 @@ class					SoundFactory
 
 private:
   bool					_exist(const std::string &);
-  Sound					*_load(const std::string &, bool, int, int);
+  Sound					*_load(const std::string &, bool, float, float);
 public:
   SoundFactory();
   ~SoundFactory();
 
-  bool					initialize(const std::string &, bool, int, int);
+  bool					initialize(const std::string &, bool, float, float);
   bool					play(const std::string &);
 };
